src/ai_companion/modules/speech/text_to_speech.py
import os
from typing import Optional

from ai_companion.core.exceptions import TextToSpeechError
from ai_companion.settings import settings
from elevenlabs import ElevenLabs, VoiceSettings


class TextToSpeech:
    """A class to handle text-to-speech operations using Elevenlabs"""

    REQUIRED_ENV_VARS=["ELEVENLABS_API_KEY", "ELEVENLABS_VOICE_ID"]

    def __init__(self):
        self._validate_env_vars()
        self._client: Optional[ElevenLabs] = None

    def validate_env_vars(self) -> None:
        """Initialize TextToSpeech class and validate environment variables"""
        missing_vars = [var for var in self.REQUIRED_ENV_VARS if not os.getenv(var)]
        if missing_vars:
            raise ValueError(f"Missing required environment variables: {', '.join(missing_vars)}")
        
    @property
    def client(self) -> ElevenLabs:
        """Get or create an Elevenlabs client instance using the singleton pattern"""
        if self._client is None:
            self._client = ElevenLabs(api_key=settings.ELEVENLABS_API_KEY)
        return self._client
    
    async def synthesize(self, text: str) -> bytes:
        """Convert text to speech using ElevenLabs
        
        Args:
            text: Text to convert to speech
        
        Returns:
            bytes: Audio data
        
        Raises:
            ValueError: If the input text is empty or too long
            TextToSpeechError: If the text-to-speech conversion fails
        """

        if not text.strip():
            raise ValueError("Input text cannot be empty.")
        
        if len(text) > 5000:
            raise ValueError("Input length exceeds maximum length of 5000 characters.")
        
        try:
            audio_generator = self.client.text_to_speech.convert(
                voice_id=settings.ELEVENLABS_VOICE_ID,
                text=text,
                model_id=settings.TTS_MODEL_NAME,
                voice_settings= VoiceSettings(
                    stability=0.5,
                    similarity_boost=0.5
                )
            )

            audio_bytes = b"".join(audio_generator)
            
            if not audio_bytes:
                raise TextToSpeechError("Generated audio is empty")
            
            return audio_bytes
        
        except Exception as e:
            raise TextToSpeechError(f"Text-to-speech conversion failed: {str(e)}") from e