pyproject.toml
[project]
name = "tela-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "absl-py==2.3.1",
    "accelerate==1.9.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "argcomplete==3.6.2",
    "asttokens==3.0.0",
    "astunparse==1.6.3",
    "attrs==25.3.0",
    "bitsandbytes==0.46.1",
    "blinker==1.9.0",
    "boto3==1.40.45",
    "botocore==1.40.45",
    "cachetools==5.5.2",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "colorama==0.4.6",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cryptography==45.0.5",
    "cycler==0.12.1",
    "dataclasses-json==0.6.7",
    "debugpy==1.8.16",
    "decorator==5.2.1",
    "distro==1.9.0",
    "executing==2.2.1",
    "faiss-cpu==1.12.0",
    "fastapi==0.116.1",
    "filelock==3.18.0",
    "filetype==1.2.0",
    "flask==3.1.1",
    "flask-cors==6.0.1",
    "flatbuffers==25.2.10",
    "fonttools==4.59.2",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "gast==0.6.0",
    "google-ai-generativelanguage>=0.6.15,<0.7.0",
    "google-api-core==2.25.0",
    "google-api-python-client==2.184.0",
    "google-auth==2.40.3",
    "google-auth-httplib2==0.2.0",
    "google-generativeai==0.8.5",
    "google-pasta==0.2.0",
    "googleapis-common-protos==1.70.0",
    "greenlet==3.2.3",
    "groq==0.31.1",
    "grpcio==1.73.0",
    "grpcio-status==1.71.2",
    "gtts==2.5.4",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "h5py==3.14.0",
    "httpcore==1.0.9",
    "httplib2==0.31.0",
    "httpx==0.28.1",
    "httpx-sse==0.4.1",
    "huggingface-hub==0.33.4",
    "idna==3.10",
    "ipykernel==6.30.1",
    "ipython==9.5.0",
    "ipython-pygments-lexers==1.1.1",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jmespath==1.0.1",
    "joblib==1.5.1",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "keras==3.11.3",
    "kiwisolver==1.4.9",
    "langchain==0.3.27",
    "langchain-community==0.3.27",
    "langchain-core==0.3.75",
    "langchain-groq==0.3.8",
    "langchain-text-splitters==0.3.9",
    "langgraph==0.4.8",
    "langgraph-checkpoint==2.0.26",
    "langgraph-prebuilt==0.2.2",
    "langgraph-sdk==0.1.70",
    "langsmith==0.3.45",
    "leap-bundle==0.5.0",
    "libclang==18.1.1",
    "lz4==4.4.4",
    "markdown==3.9",
    "markdown-it-py==4.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "matplotlib==3.10.6",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "ml-dtypes==0.5.3",
    "mpmath==1.3.0",
    "mtcnn==1.0.0",
    "multidict==6.6.4",
    "mypy-extensions==1.1.0",
    "mysql-connector-python==9.4.0",
    "namex==0.1.0",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "nltk==3.9.1",
    "numpy==2.2.6",
    "ollama==0.5.3",
    "opencv-python==4.12.0.88",
    "opt-einsum==3.4.0",
    "optree==0.17.0",
    "orjson==3.10.18",
    "ormsgpack==1.10.0",
    "packaging==24.2",
    "pandas==2.3.3",
    "parso==0.8.5",
    "pdfminer-six==20250506",
    "pdfplumber==0.11.7",
    "pillow==11.3.0",
    "pipx==1.7.1",
    "platformdirs==4.4.0",
    "prompt-toolkit==3.0.52",
    "propcache==0.3.2",
    "proto-plus==1.26.1",
    "protobuf==5.29.5",
    "psutil==7.0.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pyaudio==0.2.14",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pydub==0.25.1",
    "pygments==2.19.2",
    "pymupdf==1.26.3",
    "pyparsing==3.2.3",
    "pypdfium2==4.30.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-multipart==0.0.20",
    "pytz==2025.2",
    "pywin32==311; sys_platform == 'win32'",
    "pyyaml==6.0.2",
    "pyzmq==27.1.0",
    "regex==2024.11.6",
    "reportlab==4.4.3",
    "requests==2.32.4",
    "requests-toolbelt==1.0.0",
    "rich==14.1.0",
    "rsa==4.9.1",
    "s3transfer==0.14.0",
    "safetensors==0.5.3",
    "scikit-learn==1.7.1",
    "scipy==1.16.0",
    "sentence-transformers==5.0.0",
    "setuptools==80.9.0",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "speechrecognition==3.14.3",
    "sqlalchemy==2.0.41",
    "sqlmodel==0.0.27",
    "stack-data==0.6.3",
    "standard-aifc==3.13.0",
    "standard-chunk==3.13.0",
    "starlette==0.47.1",
    "sympy==1.14.0",
    "tavily-python==0.7.10",
    "tenacity==9.1.2",
    "tensorboard==2.20.0",
    "tensorboard-data-server==0.7.2",
    "tensorflow==2.20.0",
    "termcolor==3.1.0",
    "tf-keras==2.20.1",
    "threadpoolctl==3.6.0",
    "tiktoken==0.9.0",
    "tokenizers==0.21.2",
    "torch==2.7.1",
    "tornado==6.5.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "transformers==4.53.2",
    "typer==0.19.2",
    "typing-extensions==4.14.0",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.4.0",
    "userpath==1.9.2",
    "uvicorn==0.35.0",
    "wcwidth==0.2.13",
    "werkzeug==3.1.3",
    "wheel==0.45.1",
    "wrapt==1.17.3",
    "xxhash==3.5.0",
    "yarl==1.20.1",
    "yt-dlp==2025.9.23",
    "zstandard==0.23.0",
]

[dependency-groups]
dev = [
    "gunicorn==23.0.0",
]
